Added unit tests to ensure the read/write throughput provisioning for tables created with the DDB connector are set properly.
I've added a return to the createTable method so it now returns a
TableDescription instance, which is a Java instance.  I'm not
completely sure how I feel about that, since the goal here was to make
an adapter to CFML.  I might want to adapt this to a struct that's
created inside the connector before returning the data.