Some decent refactoring in the DynamoDB Client. This commit sees the new unit test tableShouldBeCreatedWithSpecifiedStringHashKey which needs to validate in the return of createTable() that these keys were set properly.
Turns out createTable wasn't making use of the logic within
getTableInformation, which included mapping the TableDescription AWS
instance's values into a basic struct.  That's been factored out of
getTableInformation now into a private method in the client that
createTable is also making use of in order to deliver it's own
response.  Now both methods are delivering a CFML-native struct that
describes the table and all it's relevant properties.